Relocated error and success modals
Relocated error modal and success modal pages and updated all necessary paths
